Added font size selector feature to the calculator
